Add files via upload
This is the PPT for our presentation. You can work on it and upload the updated file here itself. 
This file can be used for future team research stages. 
We can keep track on the progress we've made on the research, thus using it to demonstrate to the faculty members.